+++ AuRoRA/src/vcs/CHANGELOG.rst
@@ -6,10 +6,27 @@ Changelog for package Vital Circulatory System
 
 **Description:** This system monitors and manages the vital signs of AGi robots, including their operational status, user interactions, and network connectivity. It provides real-time feedback and alerts for critical conditions, ensuring the smooth operation of the AGi ecosystem.
 
+Future Implementation Roadmap
+-----------------------------
+* Add the following modules to VTC:
+* - Regulator: Normalizes raw sensor data into proper format and regulates the oscillation rhythm according to RTT
+* - Oscillator: Packs and encodes sensor data into vital pulse signal and publishes at oscillation rhythm
+* - Orchestrator: Monitors vital pulse signal and response, detects disconnections, and triggers safety interlocks
+
+
 Forthcoming
 -----------
 * Added launch file to launch vtc and vcc nodes based on role parameter
 * Changed the threaded terminal print to use rclpy logger
+* Added QoS Profile to vital pulse signal and response
+* Established the framework to implement PRO2 architectural for VTC:
+* - Pump: Collects glob from different lifestream flow channels
+* Decoupled the Pump Module to its own module to have better modularity
+* Optimized lifestream collection using Triple Action Pump (TAP) technique to prioritize crucial glob while conserving memory footprints
+* Implemented impurity detection and removal to ensure clean glob is harvested
+* Transitioned the naming schema to a "Biological Hardware" schema with more evocative and memorable terminologies
+* Added a standalone validation block in the pump module for real-time verification of its integrity
+
 
 0.1.2 (2026-01-01)
 ------------------